requirements.txt
setuptools
pandas
numpy
seaborn
scikit-learn
mysql-connector-python
pymysql
python-dotenv
# -e .